decouple RouteInfoFactory
