+++ packages/NetteToSymfony/config/config.yml
@@ -1,0 +1,8 @@
+services:
+    _defaults:
+        public: true
+        autowire: true
+
+    Rector\NetteToSymfony\:
+        resource: '../src'
+        exclude: '../src/{Annotation,Route/RouteInfo.php}'
