+++ packages/NetteToSymfony/tests/Rector/MethodCall/RouterListToControllerAnnotationsRetor/Fixture/general_method_named_routes.php.inc
@@ -1,0 +1,67 @@
+<?php
+
+namespace Rector\NetteToSymfony\Tests\Rector\MethodCall\RouterListToControllerAnnotationsRetor\Fixture;
+
+use Rector\NetteToSymfony\Tests\Rector\MethodCall\RouterListToControllerAnnotationsRetor\Source\Route;
+use Rector\NetteToSymfony\Tests\Rector\MethodCall\RouterListToControllerAnnotationsRetor\Source\RouteList;
+
+final class GeneralMethodNamedRoutesRouterFactory
+{
+    public function create(): RouteList
+    {
+        $routeList = new RouteList();
+        $routeList[] = new Route('<presenter>/<action>', 'Homepage:default');
+
+        return $routeList;
+    }
+}
+
+final class GeneralMethodNamedRoutesSomePresenter
+{
+    public function actionFirst()
+    {
+    }
+
+    public function actionSecond()
+    {
+    }
+}
+
+?>
+-----
+<?php
+
+namespace Rector\NetteToSymfony\Tests\Rector\MethodCall\RouterListToControllerAnnotationsRetor\Fixture;
+
+use Rector\NetteToSymfony\Tests\Rector\MethodCall\RouterListToControllerAnnotationsRetor\Source\Route;
+use Rector\NetteToSymfony\Tests\Rector\MethodCall\RouterListToControllerAnnotationsRetor\Source\RouteList;
+
+final class GeneralMethodNamedRoutesRouterFactory
+{
+    public function create(): RouteList
+    {
+        $routeList = new RouteList();
+        $routeList[] = new Route('<presenter>/<action>', 'Homepage:default');
+
+        return $routeList;
+    }
+}
+
+final class GeneralMethodNamedRoutesSomePresenter
+{
+    /**
+     * @Symfony\Component\Routing\Annotation\Route(path="general-method-named-routes-some/first")
+     */
+    public function actionFirst()
+    {
+    }
+
+    /**
+     * @Symfony\Component\Routing\Annotation\Route(path="general-method-named-routes-some/second")
+     */
+    public function actionSecond()
+    {
+    }
+}
+
+?>
